requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
executing==1.2.0
fastapi==0.95.1
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
idna==3.4
ipdb==0.13.13
ipython==8.13.2
jedi==0.18.2
matplotlib-inline==0.1.6
multidict==6.0.4
openai==0.27.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.7
Pygments==2.15.1
python-dotenv==1.0.0
PyYAML==6.0
requests==2.30.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.26.1
tomli==2.0.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
yarl==1.9.2
youtube-transcript-api==0.6.0
